Tweak command to set `GIT_AUTHOR_DATE` and `GIT_COMMITTER_DATE`
This is because the `--date` argument only sets the author date. I believe it's nicer to also have the committer date set, so all commits aren't shown with the same date in clients like Sublime Merge.

I have also added the argument `--no-gpg-sign`, as it makes the process faster and having signatures isn't useful in this case.