Fix contributions being counted multiple times by GitHub
To prevent GitHub from counting the same contribution more than once — see details in `prepareRepo`'s doc — we use the hash of each commit to identify if it's already present in the contribution repository. We do not recreate the contribution repository from scratch anymore, we keep it and only add missing contribution commits to it.